涂鸦板手机修正

+++ resources/js/app.js
@@ -20,7 +20,7 @@ axios.interceptors.response.use(
         return response
     },
     error => {
-        if (error.response.status !== undefined || error.response.status === 401) {
+        if (error.response.status !== undefined && error.response.status === 401) {
             localStorage.clear('Binggan')   //如果遇到401错误(用户未认证)，就清除Binggan和Token
             localStorage.clear('Token')
             delete axios.defaults.headers.Authorization;
